Hotfix: Fix issues from offering refactor
* Open the DB connection before attempting to create a transaction
* Don't forget to pull back the offering ID when getting meetings for an offering
* Importer doesn't need to link meetings to offering, it's implicit
