+++ src/Athena.Data/Extensions/DbConnectionExtensions.cs
@@ -1,5 +1,7 @@
 ﻿using System.Data;
+using System.Data.Common;
 using System.Threading.Tasks;
+using System.Transactions;
 using Athena.Core.Exceptions;
 using Dapper;
 using Npgsql;
@@ -23,6 +25,25 @@ namespace Athena.Data.Extensions
 
                 throw;
             }
+        }
+        
+        /// <summary>
+        /// Create a transcation scope with Async Support and enlist the specified connection with the transaction
+        /// </summary>
+        /// <param name="db">The DB Connection to enlist</param>
+        /// <returns>A transaction scope</returns>
+        public static TransactionScope CreateAsyncTransactionScope(this IDbConnection db)
+        {
+            var result = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled);
+
+            if (db.State != ConnectionState.Open)
+            {
+                db.Open();
+            }
+            
+            (db as DbConnection)?.EnlistTransaction(Transaction.Current);
+
+            return result;
         }
     }
 }
