+++ src/Athena.Importer/Program.cs
@@ -49,20 +49,11 @@ namespace Athena.Importer
 
             try
             {
-                var offerings = new OfferingImporter(
-                    uri,
-                    ApiKey,
-                    new JsonFilesystemDataProvider<Offering>(DataPath)
-                );
-                
                 await new GenericImporter<Campus>(uri, ApiKey, new JsonFilesystemDataProvider<Campus>(DataPath)).Import();
                 await new GenericImporter<Institution>(uri, ApiKey, new JsonFilesystemDataProvider<Institution>(DataPath)).Import();
                 await new GenericImporter<Course>(uri, ApiKey, new JsonFilesystemDataProvider<Course>(DataPath)).Import();
-                
-                await offerings.Import();
+                await new GenericImporter<Offering>(uri, ApiKey, new JsonFilesystemDataProvider<Offering>(DataPath)).Import();
                 await new GenericImporter<Meeting>(uri, ApiKey, new JsonFilesystemDataProvider<Meeting>(DataPath)).Import();
-                await offerings.LinkMeetings();
-                
                 await new GenericImporter<Core.Models.Program>(uri, ApiKey, new JsonFilesystemDataProvider<Core.Models.Program>(DataPath)).Import();
                 await new GenericImporter<Requirement>(uri, ApiKey, new JsonFilesystemDataProvider<Requirement>(DataPath)).Import();
